Add files via upload
Jones is developing a program that performs complex bitwise operations on three integer values: a, b, and c. The program calculates the result of the expression (a & b | c) ^ (~a & b | c) using the bitwise AND, OR, and NOT operators. 



Help Jones to develop this program.

Input format :
The input consists of three integers a, b, and c.

Output format :
The output prints the calculated result.



Refer to the sample output for formatting specifications.

Code constraints :
The given test cases fall under the following constraints:

-1000 ≤ a, b, c ≤ 1000

Sample test cases :
Input 1 :
5
3
7
Output 1 :
Result of (5 & 3 | 7) ^ (~5 & 3 | 7): 0
Input 2 :
-1
6
5
Output 2 :
Result of (-1 & 6 | 5) ^ (~-1 & 6 | 5): 2