Add files via upload
Arun is given an array of integers. His task is to find the product of elements at odd positions in the array.



Write a program to help Arun that takes an integer N as input, representing the size of the array, followed by N integers representing the elements of the array. The program should then calculate and print the product of elements at odd positions.

Input format :
The first line of input consists of an integer N, representing the number of elements in the array.

The second line consists of N space-separated integers, representing the elements of the array.

Output format :
The output prints the product of elements at odd positions. Position starts from 1.



﻿Refer to the sample output for formatting specifications.

Code constraints :
The given test cases fall under the following constraints:

1 ≤ N ≤ 15

0 ≤ array elements ≤ 100

Sample test cases :
Input 1 :
4
4 0 2 1
Output 1 :
8
Input 2 :
5
1 5 7 9 2
Output 2 :
14
Input 3 :
5
1 7 0 8 9
Output 3 :
0